separate definition from construction and merge Idator and DefinitionCheck
