+++ gradle.properties
@@ -1,3 +1,3 @@
 group=com.brightcove
-version=0.1.1-SNAPSHOT
+version=0.2.0-SNAPSHOT
 description=First class validation library for the JVM
